+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## tip
 
-* FEATURE: Add links to VMUI from the query panel
-* FEATURE: Add option to show query trace
+## [v0.1.3](https://github.com/VictoriaMetrics/grafana-datasource/releases/tag/v0.1.3)
+
+Released at 08-03-2022
+
+* FEATURE: Add links to VMUI from the query panel. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/34).
+* FEATURE: Add option to show query trace from the query panel. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/36) and [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/53).
+* FEATURE: Change license to AGPLv3. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/22).
 
 * BUGFIX: respect time filter for variables update. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/47)